+++ NEWS.md
@@ -7,6 +7,8 @@
     keyword syntax and the deprecated syntax based on the backend Stan version
     (see https://mc-stan.org/docs/reference-manual/brackets-array-syntax.html
     for details).
+  * The model components `trials()` and `offset()` now function correctly in 
+    `predict()` when they contain  response variables of the model.
 
 # dynamite 1.2.1